requirements.txt
flask
google-cloud
google-cloud-datastore
google-cloud-vision
gunicorn
